add context cache smash

+++ internal/engine/context.go
@@ -1,6 +1,7 @@
 package engine
 
 import (
+	"database/sql"
 	"fmt"
 	"strings"
 
@@ -24,17 +25,6 @@ func (q *Query) BuildContext() error {
 
 	q.QueryContext.Validator = validator
 
-	err = q.BuildTables()
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (q *Query) BuildTables() error {
-	tables := make(map[string]string)
 	connector, err := duckdb.CreateConnector()
 
 	if err != nil {
@@ -46,8 +36,42 @@ func (q *Query) BuildTables() error {
 	if err != nil {
 		return err
 	}
-
 	defer db.Close()
+
+	tables := q.generateTableDDL()
+	if err = q.dropTableIfDDLHasChanged(db, tables); err != nil {
+		return err
+	}
+	err = q.BuildTables(db, tables)
+
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (q *Query) BuildTables(db *sql.DB, tables map[string]string) error {
+	for tableName, columnString := range tables {
+
+		createTableStatement := fmt.Sprintf("create table if not exists main.%s (%s)", tableName, columnString)
+
+		hlog.Debug("Creating table in DuckDB: ", createTableStatement)
+		_, err := db.Exec(createTableStatement)
+
+		if err != nil {
+			return err
+		}
+	}
+
+	q.QueryContext.Valid = true
+
+	return nil
+}
+
+func (q *Query) generateTableDDL() map[string]string {
+	tables := make(map[string]string)
+
 	for key := range q.Main.Columns {
 		split := strings.Split(key, ".")
 		tableName := split[0]
@@ -60,12 +84,15 @@ func (q *Query) BuildTables() error {
 			}
 			duckDbDataType := duckdb.PgTypeMap[sourceDataType.MajorityType]
 			if len(tables[tableName]) == 0 {
-				tables[tableName] += "hypha_source_name string"
+				tables[tableName] += "hypha_source_name VARCHAR"
 			}
 			tables[tableName] += ", " + columnName + " " + duckDbDataType
 		}
 	}
+	return tables
+}
 
+func (q *Query) dropTableIfDDLHasChanged(db *sql.DB, tables map[string]string) error {
 	for tableName, columnString := range tables {
 		var tableExists bool
 		for table := range q.Cfg.Tables {
@@ -78,23 +105,36 @@ func (q *Query) BuildTables() error {
 			continue
 		}
 
-		dropTableStatement := fmt.Sprintf("drop table if exists main.%s", tableName)
-		createTableStatement := fmt.Sprintf("create table if not exists main.%s (%s)", tableName, columnString)
-		hlog.Debug("Dropping table in DuckDB: ", dropTableStatement)
-		_, err := db.Exec(dropTableStatement)
-
+		rows, err := db.Query(fmt.Sprintf("select column_name, data_type from information_schema.columns where table_name = '%s'", tableName))
 		if err != nil {
+			hlog.Debug("Error querying information_schema.columns: ", err)
 			return err
 		}
-		hlog.Debug("Creating table in DuckDB: ", createTableStatement)
-		_, err = db.Exec(createTableStatement)
+		defer rows.Close()
+		var columns []string
+		for rows.Next() {
+			var columnName, dataType string
+			err = rows.Scan(&columnName, &dataType)
+			if err != nil {
+				hlog.Debug("Error scanning rows: ", err)
+				return err
+			}
+			columns = append(columns, columnName+" "+dataType)
+		}
 
-		if err != nil {
-			return err
+		columnParts := strings.Split(columnString, ",")
+
+		for _, column := range columns {
+			if !strings.Contains(columnString, column) || len(columnParts) != len(columns) {
+				hlog.Debug("Table DDL has changed for table ", tableName)
+				dropTableStatement := fmt.Sprintf("drop table if exists main.%s", tableName)
+				hlog.Debug("Dropping table in DuckDB: ", dropTableStatement)
+				_, err := db.Exec(dropTableStatement)
+
+				return err
+			}
 		}
 	}
-
-	q.QueryContext.Valid = true
 
 	return nil
 }
